Fix C4146 unary minus operator applied to unsigned type (#4065)
* Fix C4146 unary minus operator applied to unsigned type
* Silence C4267 'argument' conversion from 'size_t' to X

+++ SPIRV/SpvBuilder.cpp
@@ -1418,7 +1418,7 @@ Id Builder::createDebugLocalVariable(Id type, char const*const name, size_t cons
     inst->addIdOperand(currentDebugScopeId.top()); // scope id
     inst->addIdOperand(makeUintConstant(NonSemanticShaderDebugInfo100FlagIsLocal)); // flags id
     if(argNumber != 0) {
-        inst->addIdOperand(makeUintConstant(argNumber));
+        inst->addIdOperand(makeUintConstant(static_cast<unsigned int>(argNumber)));
     }
 
     constantsTypesGlobals.push_back(std::unique_ptr<Instruction>(inst));
